RD-4142: Refinement of DeploymentList widget
