+++ widgets/deploymentList/src/types.ts
@@ -1,0 +1,19 @@
+export declare namespace DeploymentListWidget {
+    export type Configuration = {
+        pageSize: number;
+        sortColumn: string;
+        sortAscending: boolean;
+    };
+}
+
+type DeploymentsTableItemType = {
+    id: string;
+    displayName: string;
+    blueprintId: string;
+    label: string;
+};
+
+export type DeploymentsTableDataType = {
+    total: number;
+    items: DeploymentsTableItemType[];
+};
