--- widgets/deploymentList/src/types/DeploymentList.ts
@@ -1,7 +1,0 @@
-export declare namespace DeploymentListWidget {
-    export type Configuration = {
-        pageSize: number;
-        sortColumn: string;
-        sortAscending: boolean;
-    };
-}
