--- widgets/deploymentList/src/types/DeploymentsTableDataType.ts
@@ -1,13 +1,0 @@
-type DeploymentsTableItemType = {
-    id: string;
-    displayName: string;
-    blueprintId: string;
-    label: string;
-};
-
-type DeploymentsTableDataType = {
-    total: number;
-    items: DeploymentsTableItemType[];
-};
-
-export default DeploymentsTableDataType;
